pyproject.toml
[tool.poetry]
name = "codegrapher"
version = "0.1.0"
description = ""
authors = ["Arslan Ahmad <arslanahmadpraimy786@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
uvicorn = "^0.29.0"
pymongo = "^4.7.2"
motor = "^3.4.0"
pydantic = "^2.7.1"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pyjwt = "^2.8.0"
python-dotenv = "^1.0.1"
aiofiles = "^23.2.1"
celery = "^5.4.0"
pre-commit = "^3.7.1"
redis = "^5.0.4"
pytest = "^8.2.1"
pytest-asyncio = "^0.23.7"
sentry-sdk = {extras = ["fastapi"], version = "^2.2.1"}
httpx = "^0.27.0"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.7.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
